Completed Appointment_procedure, Invoice, Display Reviews
Trying to get dentist names right now, specifically dentists who have done a procedure on the current patient. Also, trying to make a dropdown menu so that the patient can submit a review. The dropdown menu also needs to fetch ONLY THE NAMES OF THE DENTISTS WHO HAVE DONE A PROCEDURE ON THE PATIENT. It's not supposed to fetch ALL dentists.