Completed patient_landing.php (see description)
Review box is fixed completed. Representative now shows an unordered list. Icons were added to section headers. Added review description parameter to Review. The only problems left are for enhancement purposes, but
are not required. Those are ,, show only 10 rows at a time per section, then have a Newer/Older navigation at the bottom ,, Sort each box by date, ID, etc. ,, Make it so that the Submit button in review does not reload the page to the top. Other than that, it's done.